FAD-4498 Move NodeLink component & Fix a small scroll issue
